tests 4 and 5 functional.

+++ src/CineReviewsPackage/Shows/ShowComparator.java
@@ -5,9 +5,9 @@ import java.util.Comparator;
 public class ShowComparator implements Comparator<Show> {
     @Override
     public int compare(Show show1, Show show2) {
-        int scoreComparison = show1.getAverageReviews() - show2.getAverageReviews();
+        double scoreComparison = show1.getAverageReviews() - show2.getAverageReviews();
         if (scoreComparison != 0) {
-            return scoreComparison;
+            return (int)scoreComparison;
         }
 
         int yearComparison = Integer.compare(show1.getYear(), show2.getYear());
